+++ examples/tpch/queries/Q14.h
@@ -6,12 +6,12 @@ double Q14(maps m, const char* start_date, const char* end_date) {
   
   ship_map s_range = ship_map::range(m.shipments_for_date, Date(start_date), Date(end_date));
 
-  using sum = pair<float,float>;
-  using Add = Add_Pair<Add<float>,Add<float>>;
+  using sum = pair<double,double>;
+  using Add = Add_Pair<Add<double>,Add<double>>;
 
   auto date_f = [&] (ship_map::E& e) -> sum {
     auto li_f = [&] (line_item_set::E& l) -> sum {
-      float dp = l.e_price * (1.0 - l.discount.val());
+      double dp = l.e_price * (1.0 - l.discount.val());
 	  char* x = static_data.all_part[l.part_key].type();
       if (memcmp(x, promo, plen) == 0) return sum(dp,dp);
       else return sum(dp,0.0);
